+++ MatrixSDK/Data/MXRoom.h
@@ -36,6 +36,7 @@
 #import "MXEventTimeline.h"
 #import "MXEventsEnumerator.h"
 #import "MXCryptoConstants.h"
+#import "MXSendReplyEventStringsLocalizable.h"
 
 @class MXRoom;
 @class MXSession;
@@ -753,6 +754,40 @@ FOUNDATION_EXPORT NSString *const kMXRoomDidFlushDataNotification;
 - (MXHTTPOperation*)setRelatedGroups:(NSArray<NSString *>*)relatedGroups
                              success:(void (^)(void))success
                              failure:(void (^)(NSError *error))failure;
+
+
+/**
+ Indicate if replying to the provided event is supported.
+ Only event of type 'MXEventTypeRoomMessage' are supported for the moment, and for certain msgtype.
+ 
+ @param eventToReply the event to reply to
+ @return YES if it is possible to reply to this event
+ */
+- (BOOL)canReplyToEvent:(MXEvent *)eventToReply;
+
+/**
+ Send a reply to an event with text message to the room.
+ 
+ It's only supported to reply to event with 'm.room.message' event type and following message types: 'm.text', 'm.text', 'm.emote', 'm.notice', 'm.image', 'm.file', 'm.video', 'm.audio'.
+ 
+ @param eventToReply The event to reply.
+ @param textMessage the text to send.
+ @param formattedTextMessage the optional HTML formatted string of the text to send.
+ @param stringLocalizations string localizations used when building reply message.
+ @param localEcho a pointer to a MXEvent object (@see sendMessageWithContent: for details).
+ @param success A block object called when the operation succeeds. It returns
+ the event id of the event generated on the home server
+ @param failure A block object called when the operation fails.
+ 
+ @return a MXHTTPOperation instance.
+ */
+- (MXHTTPOperation*)sendReplyToEvent:(MXEvent*)eventToReply
+                     withTextMessage:(NSString*)textMessage
+                formattedTextMessage:(NSString*)formattedTextMessage
+                 stringLocalizations:(id<MXSendReplyEventStringsLocalizable>)stringLocalizations
+                           localEcho:(MXEvent**)localEcho
+                             success:(void (^)(NSString *eventId))success
+                             failure:(void (^)(NSError *error))failure NS_REFINED_FOR_SWIFT;
 
 
 #pragma mark - Events timeline