+++ CHANGES.rst
@@ -7,6 +7,7 @@ Improvements:
 * MXRoomSummary: Add a membersCount property to cache MXRoomState one
 * MXRoomSummary: Add a membership property to cache MXRoomState one
 * MXStore: Add Obj-C annotations.
+* MXRoom: Add send reply with text message (vector-im/riot-ios#1911).
 
 Bug fix: